+++ Unity Project/Assets/Scripts/PlayerInput.cs
@@ -2,38 +2,18 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class PlayerInput : MonoBehaviour
+namespace Game
 {
-    public CharacterController controller;
-
-    public void Update()
+    public class PlayerInput : MonoBehaviour
     {
-        float horizontal = Input.GetAxisRaw("Horizontal");
-        float vertical = Input.GetAxisRaw("Vertical");
-        Vector3 direction = new Vector3(horizontal, 0f, vertical).normalized;
-    }
-}
-
-public class PlayerController : MonoBehaviour
-{
-    public float speed = 1f;
-    public float turnSmoothTime = 0.1f;
-    public Animator animator;
-
-    new public Transform camera;
-    public float turnSmoothVelocity;
-
-    public void Update()
-    {
-        var direction = PlayerInput.GetDirection();
-        if (direction.magnitude >= 0.1f)
+        public Vector3 GetDirection()
         {
-            float targetAngle = Mathf.Atan2(direction.x, direction.z) * Mathf.Rad2Deg + camera.eulerAngles.y;
-            float angle = Mathf.SmoothDampAngle(transform.eulerAngles.y, targetAngle, ref turnSmoothVelocity, turnSmoothTime);
-            transform.rotation = Quaternion.Euler(0f, angle, 0f);
-
-            Vector3 moveDir = Quaternion.Euler(0f, targetAngle, 0f) * Vector3.forward;
-            controller.Move(moveDir.normalized * speed * Time.deltaTime);
+            float horizontal = Input.GetAxisRaw("Horizontal");
+            float vertical = Input.GetAxisRaw("Vertical");
+            
+            return new Vector3(horizontal, 0f, vertical).normalized;
         }
     }
 }
+
+
